server/elerning/requirements.txt
Django==4.2.6
django-ckeditor==6.7.0
django-cors-headers==4.3.0
django-debug-toolbar==4.2.0
django-environ==0.11.2
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
djoser==2.2.0
drf-yasg==1.21.7
flake8==6.1.0
Markdown==3.4.4
oauthlib==3.2.2
Pillow==10.1.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pycparser==2.21
pyflakes==3.1.0
PyJWT==2.8.0
python-magic==0.4.27
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1